+++ lib/hadith_nawawi.dart
@@ -1,4 +1,13 @@
 library hadith_nawawi;
 
+// Export models
 export 'src/model/hadith_model.dart';
+
+// Export services
 export 'src/services/hadith_service.dart';
+
+// Export widgets
+export 'src/widgets/hadith_card.dart';
+export 'src/widgets/hadith_detail_view.dart';
+export 'src/widgets/hadith_list_view.dart';
+export 'src/widgets/hadith_search_bar.dart';